Dissolve CssSerializer into the Css Showable and Streamable instances
`CssSerializer` was a public object whose `render`/`emit` duplicated what
the `Css is Showable` and `Css is Streamable by Text` instances already
expose via `.show` / streaming. The collect and lazy-producer driving now
live in those two givens; the shared serialization fold is a private
`write` method on `object Css`, so the typeclasses are the single public
route. The graffiti stylesheet helper and the cataclysm tests move to
`.show` (with the chosen `Css.Formatting` brought into scope by import).
