+++ lib/cataclysm/src/core/soundness_cataclysm_core.scala
@@ -34,7 +34,7 @@ package soundness
 
 export cataclysm.{Css, CssError, CssErrors, cssAggregable, SelectorList, Selector, Compound,
     Simple, Combinator, AttributeMatcher, AttributeTest, Prefix, PseudoArgument,
-    CssSerializer, CssConvertible, Pixels, Rems, Exs, Chs, ViewportWidths,
+    CssConvertible, Pixels, Rems, Exs, Chs, ViewportWidths,
     ViewportHeights, ViewportMins, ViewportMaxes, Percents, Degrees, Radians, Turns, Flexes, Px,
     Rem, Ex, Ch, Vw, Vh, Vmin, Vmax, Cm, Mm, Pt, Pc, Pct, S, Ms, Deg, Rad, Turn, Fr, css}
